Saving cell types for scRNAseq data
